requirements.txt
torch
torch-optimizer
numpy
cython
flatland-rl
gym
tqdm